* Ajustado a precisão de valores das classes com BigDecimal * Ajustado a forma como é apresentado a regra dos estabelecimentos utilizando o padrão Strategy. * Adicionado uma interface das classes Bissness para evitar a quebra do 4 principio do SOLID.
