--- src/main/java/br/com/alelo/consumer/consumerpat/ConsumerTestApplication.java
@@ -1,13 +1,0 @@
-package br.com.alelo.consumer.consumerpat;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ConsumerTestApplication {
-
-	public static void main(String... args) {
-		SpringApplication.run(ConsumerTestApplication.class, args);
-	}
-
-}
