--- src/main/java/br/com/alelo/consumer/consumerpat/constant/BalanceCardStandard.java
@@ -1,17 +1,0 @@
-package br.com.alelo.consumer.consumerpat.constant;
-
-import lombok.Getter;
-
-@Getter
-public enum BalanceCardStandard {
-
-    ESTABLISHMENT_NAME("Recarga"),
-    PRODUCT_DESCRIPTION("Recarga do cartão"),;
-
-    public final String description;
-
-    BalanceCardStandard(String description) {
-        this.description = description;
-    }
-
-}
