--- src/main/java/br/com/alelo/consumer/consumerpat/constant/DocumentNumberValidator.java
@@ -1,110 +1,0 @@
-package br.com.alelo.consumer.consumerpat.constant;
-
-import java.util.InputMismatchException;
-
-public class DocumentNumberValidator {
-
-    public static boolean verify(String cnp) {
-        cnp = cnp.replaceAll("[^0-9]", "");
-
-        if (cnp.length() == 11) {
-            return toCpf(cnp);
-        } else if (cnp.length() == 14) {
-            return toCnpj(cnp);
-        } else {
-            return false;
-        }
-    }
-
-    private static boolean toCpf(String cpf) {
-
-        cpf = cpf.replaceAll("[^0-9]", "");
-
-        if (cpf.length() != 11 || cpf.matches("(\\d)\\1{10}")) {
-            return false;
-        }
-
-        int[] digits = new int[11];
-        for (int i = 0; i < 11; i++) {
-            digits[i] = cpf.charAt(i) - '0';
-        }
-        int sum = 0;
-        for (int i = 0; i < 9; i++) {
-            sum += digits[i] * (10 - i);
-        }
-        int remainder = sum % 11;
-        int digit1 = (remainder < 2) ? 0 : (11 - remainder);
-
-
-        sum = 0;
-        for (int i = 0; i < 10; i++) {
-            sum += digits[i] * (11 - i);
-        }
-        remainder = sum % 11;
-        int digit2 = (remainder < 2) ? 0 : (11 - remainder);
-
-
-        return (digit1 == digits[9] && digit2 == digits[10]);
-    }
-
-    private static boolean toCnpj(String cnpj) {
-        cnpj = cnpj.replace(".", "");
-        cnpj = cnpj.replace("-", "");
-        cnpj = cnpj.replace("/", "");
-
-        try{
-            Long.parseLong(cnpj);
-        } catch(NumberFormatException e){
-            return false;
-        }
-
-        // considera-se erro CNPJ's formados por uma sequencia de numeros iguais
-        if (cnpj.equals("00000000000000") || cnpj.equals("11111111111111")
-                || cnpj.equals("22222222222222") || cnpj.equals("33333333333333")
-                || cnpj.equals("44444444444444") || cnpj.equals("55555555555555")
-                || cnpj.equals("66666666666666") || cnpj.equals("77777777777777")
-                || cnpj.equals("88888888888888") || cnpj.equals("99999999999999") || (cnpj.length() != 14))
-            return (false);
-        char dig13, dig14;
-        int sm, i, r, num, peso;
-        try {
-            sm = 0;
-            peso = 2;
-            for (i = 11; i >= 0; i--) {
-                num = (int) (cnpj.charAt(i) - 48);
-                sm = sm + (num * peso);
-                peso = peso + 1;
-                if (peso == 10)
-                    peso = 2;
-            }
-
-            r = sm % 11;
-            if ((r == 0) || (r == 1))
-                dig13 = '0';
-            else
-                dig13 = (char) ((11 - r) + 48);
-
-            sm = 0;
-            peso = 2;
-            for (i = 12; i >= 0; i--) {
-                num = (int) (cnpj.charAt(i) - 48);
-                sm = sm + (num * peso);
-                peso = peso + 1;
-                if (peso == 10)
-                    peso = 2;
-            }
-            r = sm % 11;
-            if ((r == 0) || (r == 1))
-                dig14 = '0';
-            else
-                dig14 = (char) ((11 - r) + 48);
-            if ((dig13 == cnpj.charAt(12)) && (dig14 == cnpj.charAt(13)))
-                return (true);
-            else
-                return (false);
-        } catch (InputMismatchException erro) {
-            return (false);
-        }
-    }
-
-}
