--- src/main/java/br/com/alelo/consumer/consumerpat/dto/BuyRequest.java
@@ -1,21 +1,0 @@
-package br.com.alelo.consumer.consumerpat.dto;
-
-import br.com.alelo.consumer.consumerpat.constant.EstablishmentType;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
-
-@Data
-public class BuyRequest {
-
-    @JsonProperty
-    private EstablishmentType establishmentType;
-    @JsonProperty
-    private String establishmentName;
-    @JsonProperty
-    private String cardNumber;
-    @JsonProperty
-    private String productDescription;
-    @JsonProperty
-    private Double value;
-
-}
