--- src/main/java/br/com/alelo/consumer/consumerpat/exception/AddConsumerCardExeption.java
@@ -1,7 +1,0 @@
-package br.com.alelo.consumer.consumerpat.exception;
-
-public class AddConsumerCardExeption extends RuntimeException{
-    public AddConsumerCardExeption(String message) {
-        super(message);
-    }
-}
