--- src/main/java/br/com/alelo/consumer/consumerpat/exception/BuyException.java
@@ -1,7 +1,0 @@
-package br.com.alelo.consumer.consumerpat.exception;
-
-public class BuyException extends RuntimeException {
-    public BuyException(String message) {
-        super(message);
-    }
-}
