--- src/main/java/br/com/alelo/consumer/consumerpat/exception/ConsumerNotFoundException.java
@@ -1,7 +1,0 @@
-package br.com.alelo.consumer.consumerpat.exception;
-
-public class ConsumerNotFoundException extends RuntimeException{
-    public ConsumerNotFoundException(String message) {
-        super(message);
-    }
-}
