--- src/main/java/br/com/alelo/consumer/consumerpat/exception/SaveConsumerException.java
@@ -1,8 +1,0 @@
-package br.com.alelo.consumer.consumerpat.exception;
-
-public class SaveConsumerException extends RuntimeException{
-
-        public SaveConsumerException(String message) {
-            super(message);
-        }
-}
