--- src/main/java/br/com/alelo/consumer/consumerpat/exception/SetBalanceCardExeption.java
@@ -1,7 +1,0 @@
-package br.com.alelo.consumer.consumerpat.exception;
-
-public class SetBalanceCardExeption extends RuntimeException{
-    public SetBalanceCardExeption(String message) {
-        super(message);
-    }
-}
