--- src/main/java/br/com/alelo/consumer/consumerpat/exception/standard/StandardResponseError.java
@@ -1,19 +1,0 @@
-package br.com.alelo.consumer.consumerpat.exception.standard;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.Data;
-
-import java.util.Date;
-
-@Data
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class StandardResponseError {
-    private Date timestamp;
-    private String method;
-    private String message;
-    private Object errors;
-    private String exception;
-    private String status;
-    private String path;
-    private Boolean error;
-}
