--- src/main/java/br/com/alelo/consumer/consumerpat/repository/ConsumerRepository.java
@@ -1,12 +1,0 @@
-package br.com.alelo.consumer.consumerpat.repository;
-
-import br.com.alelo.consumer.consumerpat.entity.Consumer;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface ConsumerRepository extends JpaRepository<Consumer, Integer> {
-
-    Optional<Consumer> findConsumerByDocumentNumber(String documentNumber);
-
-}
