--- src/main/resources/config/application.yml
@@ -1,13 +1,0 @@
-spring:
-  datasource:
-    driver-class-name: org.h2.Driver
-    url: jdbc:h2:mem:consumer-pat;DB_CLOSE_DELAY=-1
-    username: sa
-    password: sa
-  jpa:
-    database-platform: org.hibernate.dialect.H2Dialect
-    hibernate:
-      ddl-auto: create-drop
-  h2:
-    console.enabled: true
-    console.path: /h2-console
